+++ @vue-start/element-pro/src/index.ts
@@ -16,6 +16,9 @@ import {
   ProShowDate,
   ProShowOptions,
   ProShowTree,
+  ProPage,
+  ProTip,
+  ProOperate,
 } from "@vue-start/pro";
 import {
   ElForm,
@@ -26,6 +29,8 @@ import {
   ElButton,
   ElDescriptionsItem,
   ElCheckbox,
+  ElTooltip,
+  ElScrollbar,
 } from "element-plus";
 import {
   FormMethods,
@@ -44,7 +49,16 @@ import {
   ProFormCascader,
 } from "./form";
 import { ProTable as Table, ProTableOperateItem } from "./table";
-import { ProUploader, ProModal as Modal, ProLoading, ProMenus, ProModal, ProPagination, ProPopover } from "./comp";
+import {
+  ProUploader,
+  ProModal as Modal,
+  ProLoading,
+  ProMenus,
+  ProModal,
+  ProPagination,
+  ProPopover,
+  ProDropdown,
+} from "./comp";
 
 export * from "./form";
 export * from "./table";
@@ -99,6 +113,12 @@ ProUploaderText.props = {
   modelValue: String,
 };
 
+ProTip.props = {
+  ...ElTooltip.props,
+  ...ProTip.props,
+  placement: { type: String, default: "top-start" },
+};
+
 export const elementMap = {
   //element-plus
   [ElementKeys.LoadingKey]: ProLoading,
@@ -111,18 +131,23 @@ export const elementMap = {
   [ElementKeys.ModalKey]: ProModal,
   [ElementKeys.PaginationKey]: ProPagination,
   [ElementKeys.PopoverKey]: ProPopover,
+  [ElementKeys.TooltipKey]: ElTooltip,
   [ElementKeys.CheckboxKey]: ElCheckbox,
   [ElementKeys.FormKey]: Form,
   [ElementKeys.FormItemKey]: ProFormItem,
   [ElementKeys.TableKey]: Table,
   [ElementKeys.TableOperateKey]: ProTableOperateItem,
   [ElementKeys.UploaderKey]: ProUploader,
+  [ElementKeys.DropdownKey]: ProDropdown,
+  [ElementKeys.ScrollKey]: ElScrollbar,
   //pro
   [ElementKeys.ProFormKey]: ProForm,
   [ElementKeys.ProSearchFormKey]: ProSearchForm,
   [ElementKeys.ProTableKey]: ProTable,
   [ElementKeys.ProCurdKey]: ProCurd,
   [ElementKeys.ProListKey]: ProList,
+  [ElementKeys.ProPageKey]: ProPage,
+  [ElementKeys.ProOperateKey]: ProOperate,
   //show
   text: ProShowText,
   digit: ProShowDigit,