+++ src/layout/index.tsx
@@ -1,13 +1,12 @@
 import { defineComponent, ref } from "vue";
-import { RouterView } from "vue-router";
-import { findTreeItem, useEffect } from "@vue-start/hooks";
+import { useEffect } from "@vue-start/hooks";
 import { ProLayout, useProRouter } from "@vue-start/pro";
-import { find } from "lodash";
 import { HeaderLeft, HeaderRight } from "@/layout/Header";
 import { css } from "@emotion/css";
-import { routes } from "@/router/routes";
 import { menus } from "@/common/menus";
 import { useConfigStore } from "@/store/StoreCurrent";
+// @ts-ignore
+import Sortable from "sortablejs";
 
 export const BasicLayout = defineComponent(() => {
   const { router } = useProRouter();
@@ -30,28 +29,6 @@ export const BasicLayout = defineComponent(() => {
       router.replace({ name: "OverviewIndexMd" });
     }
   }, []);
-
-  const findCurrentTopName = (route: any, menuTopMap: any) => {
-    const list = findTreeItem(routes, (item) => item.name === route.name).list;
-    if (list) {
-      const index = find(list, (item) => {
-        //根据当前项目的规则制定
-        return item.path === "index" || item.path === "base";
-      });
-      if (index) return menuTopMap[index.name];
-    }
-  };
-
-  const findActiveKey = (route: any) => {
-    const list = findTreeItem(routes, (item) => item.name === route.name).list;
-    if (list) {
-      const index = find(list, (item) => {
-        //根据当前项目的规则制定
-        return item.path === "index" || item.path === "base";
-      });
-      if (index) return index.name;
-    }
-  };
 
   const collapseRef = ref(false);
 
@@ -87,6 +64,23 @@ export const BasicLayout = defineComponent(() => {
     },
   };
 
+  let sortable: any;
+
+  const onDragRegister = ({ dom, dataIdAttr, onDragEnd }: any) => {
+    if (sortable) {
+      sortable.destroy();
+      sortable = null;
+    }
+    sortable = Sortable.create(dom, {
+      animation: 300,
+      dataIdAttr,
+      disabled: false,
+      onEnd: () => {
+        onDragEnd(sortable.toArray());
+      },
+    });
+  };
+
   return () => {
     return (
       <ProLayout
@@ -96,15 +90,13 @@ export const BasicLayout = defineComponent(() => {
           },
         })}
         layout={config.layout as any}
-        menus={menus}
-        fieldNames={{ value: "name", label: "title", children: "children" }}
-        findCurrentTopName={findCurrentTopName}
-        findActiveKey={findActiveKey}
-        menuProps={{
-          class: collapseRef.value ? "pro-layout-menus mini" : "pro-layout-menus",
-          collapse: collapseRef.value, //el
-          inlineCollapsed: collapseRef.value, //ant
+        tabs={{
+          onDragRegister,
         }}
+        menus={menus as any}
+        fieldNames={{ value: "name", label: "title", hide: "hide", children: "children" }}
+        collapse={collapseRef.value}
+        watermark={{}}
         v-slots={{
           "header-start": () => <HeaderLeft />,
           "header-end": () => (
@@ -125,9 +117,8 @@ export const BasicLayout = defineComponent(() => {
               {collapseRef.value ? "展开" : "合并"}
             </div>
           ),
-        }}>
-        <RouterView />
-      </ProLayout>
+        }}
+      />
     );
   };
 });