+++ src/platform-element-plus/index.ts
@@ -6,10 +6,10 @@ import { createStore } from "@/store";
 import { createConfig } from "@vue-start/config";
 import { createRequest, ContentTypeInterceptor } from "@vue-start/request";
 
+import '@vue-start/css/normalize.css'
 import "@vue-start/pro/index.css";
 import "@vue-start/media/index.css";
 
-import "@/style/normalize.css";
 import "@/style/index.css";
 //
 import "highlight.js/styles/github.css";
@@ -22,7 +22,7 @@ import ElementPlus from "element-plus";
 import * as ElementPlusIconsVue from '@element-plus/icons-vue'
 import "element-plus/dist/index.css";
 import 'element-plus/theme-chalk/dark/css-vars.css'
-import locale from "element-plus/lib/locale/lang/zh-cn";
+import locale from "element-plus/es/locale/lang/zh-cn";
 
 import { initComp } from "@/platform-element-plus/component";
 import { startsWith } from "lodash";